Change future reward function, remove draw bonus
The future reward now doesn't look two moves ahead for the current
player; instead, the best move of the opponent is taken into account as
a "malus".

Now we need to remove the points for drawing, otherwise making a play
that will lead the opponent to draw would be worth -50 points.  At this
point we would have accumulated ~2 points for moving the game forward,
either -100 for the opponent winning or -50 for the opponent drawing
wipes out all of our score, making us not have a preference for either.

The bot now achieves a 100% win or draw rate against the random player.

(Also small refactoring of accessing the Q table, to make it more
readable).